Merge all features into a single stable version
Implement the in-app Edu book content
Story Points: 8
Description:
The user should be able to access the book content in the app.

Implement the Drag & Drop Game in the App
Story Points: 8
Description:
The user needs to drag and drop the planets into their correct spots on the solar system.